refactor(auth): remove duplicate code
