+++ src/main/java/nextstep/subway/auth/ui/interceptor/authentication/AuthenticationInterceptor.java
@@ -1,6 +1,10 @@
 package nextstep.subway.auth.ui.interceptor.authentication;
 
 import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.domain.AuthenticationToken;
+import nextstep.subway.auth.ui.interceptor.convert.AuthenticationConverter;
+import nextstep.subway.member.application.UserDetailsService;
+import nextstep.subway.member.domain.LoginMember;
 import org.springframework.web.servlet.HandlerInterceptor;
 
 import javax.servlet.http.HttpServletRequest;
@@ -8,6 +12,42 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public abstract class AuthenticationInterceptor implements HandlerInterceptor {
+    private final AuthenticationConverter authenticationConverter;
+    private final UserDetailsService userDetailsService;
+
+    public AuthenticationInterceptor(AuthenticationConverter authenticationConverter, UserDetailsService userDetailsService) {
+        this.authenticationConverter = authenticationConverter;
+        this.userDetailsService = userDetailsService;
+    }
 
     public abstract void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException;
+
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
+        AuthenticationToken token = authenticationConverter.convert(request);
+        Authentication authentication = authenticate(token);
+
+        if (authentication == null) {
+            throw new RuntimeException();
+        }
+
+        afterAuthentication(request, response, authentication);
+        return false;
+    }
+
+    public Authentication authenticate(AuthenticationToken token) {
+        String principal = token.getPrincipal();
+        LoginMember userDetails = userDetailsService.loadUserByUsername(principal);
+        checkAuthentication(userDetails, token);
+        return new Authentication(userDetails);
+    }
+
+    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+        if (userDetails == null) {
+            throw new RuntimeException();
+        }
+
+        if (!userDetails.checkPassword(token.getCredentials())) {
+            throw new RuntimeException();
+        }
+    }
 }
