+++ src/main/java/nextstep/subway/auth/ui/interceptor/authentication/TokenAuthenticationInterceptor.java
@@ -15,29 +15,13 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class TokenAuthenticationInterceptor extends AuthenticationInterceptor {
-    private final UserDetailsService userDetailsService;
     private final JwtTokenProvider jwtTokenProvider;
     private final ObjectMapper objectMapper;
-    private final AuthenticationConverter authenticationConverter;
 
     public TokenAuthenticationInterceptor(UserDetailsService userDetailsService, JwtTokenProvider jwtTokenProvider, AuthenticationConverter authenticationConverter) {
-        this.userDetailsService = userDetailsService;
+        super(authenticationConverter, userDetailsService);
         this.jwtTokenProvider = jwtTokenProvider;
         this.objectMapper = new ObjectMapper();
-        this.authenticationConverter = authenticationConverter;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        AuthenticationToken authenticationToken = authenticationConverter.convert(request);
-        Authentication authentication = authenticate(authenticationToken);
-
-        if (authentication == null) {
-            throw new RuntimeException();
-        }
-
-        afterAuthentication(request, response, authentication);
-        return false;
     }
 
     public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
@@ -47,22 +31,5 @@ public class TokenAuthenticationInterceptor extends AuthenticationInterceptor {
         response.setStatus(HttpServletResponse.SC_OK);
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.getWriter().write(objectMapper.writeValueAsString(new TokenResponse(token)));
-    }
-
-    private Authentication authenticate(AuthenticationToken token) {
-        String principal = token.getPrincipal();
-        LoginMember userDetails = userDetailsService.loadUserByUsername(principal);
-        checkAuthentication(userDetails, token);
-        return new Authentication(userDetails);
-    }
-
-    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new RuntimeException();
-        }
-
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new RuntimeException();
-        }
     }
 }
